Do not copy template files, but re-export content instead
This has two advantages: We can take advantage of hot reload as we 
edit/modify the templates and we have less code duplication (which 
should lead to smaller bundle sizes).
