+++ templates/index.getStaticProps.js
@@ -1,0 +1,11 @@
+import getOrganizations from "helpers/getOrganizations";
+
+export function getStaticProps() {
+  const organizations = getOrganizations();
+
+  return {
+    props: {
+      organizationsCount: organizations.length,
+    },
+  };
+}
