+++ templates/map.getStaticProps.js
@@ -1,0 +1,16 @@
+import getOrganizations from "helpers/getOrganizations";
+
+export function getStaticProps() {
+  let organizations = getOrganizations();
+
+  // Keep only organizations with defined latitude and longitude
+  organizations = organizations.filter(
+    (org) => (org.latitude != null) & (org.longitude != null)
+  );
+
+  return {
+    props: {
+      organizations,
+    },
+  };
+}
