+++ templates/organizations/[slug].getStaticProps.js
@@ -1,0 +1,24 @@
+import getOrganizations from "helpers/getOrganizations";
+
+export function getStaticPaths() {
+  const organizations = getOrganizations();
+  const slugs = organizations.map((organization) => organization.slug);
+
+  return {
+    paths: slugs.map((slug) => ({
+      params: { slug },
+    })),
+    fallback: false,
+  };
+}
+
+export function getStaticProps({ params }) {
+  const { slug } = params;
+
+  const organizations = getOrganizations();
+  const organization = organizations.find((org) => org.slug === slug);
+
+  return {
+    props: { organization },
+  };
+}
