+++ templates/organizations/[slug].js
@@ -1,5 +1,3 @@
-import getOrganizations from "helpers/getOrganizations";
-
 import {
   Box,
   Card,
@@ -170,28 +168,5 @@ const OrganizationPage = ({ organization }) => {
     </Layout>
   );
 };
-
-export function getStaticPaths() {
-  const organizations = getOrganizations();
-  const slugs = organizations.map((organization) => organization.slug);
-
-  return {
-    paths: slugs.map((slug) => ({
-      params: { slug },
-    })),
-    fallback: false,
-  };
-}
-
-export function getStaticProps({ params }) {
-  const { slug } = params;
-
-  const organizations = getOrganizations();
-  const organization = organizations.find((org) => org.slug === slug);
-
-  return {
-    props: { organization },
-  };
-}
 
 export default OrganizationPage;
